test(core,media-glue): extract shared tests/common fixtures
Rust compiles each file under `tests/` as its own binary, so the
integration tests had grown copy-pasted fixtures with real drift:
`echo_subprotocol`, the "ack start then idle" WS server, the INVITE
builder, the single-route dialplan, and the SDP offer constants were
duplicated across `acceptor_metrics`, `acceptor_webhooks`,
`acceptor_cdr`, `hold_resume`, and `registry_bye` — with the INVITE
builder and route helper having diverged in signature.

Add a `tests/common/mod.rs` per crate:

- core: `echo_subprotocol`, `server_acks_start_then_idles`,
  `invite(body, uri, call_id)`, `one_route(name, ws_url)`, and the
  `LINPHONE_PCMU_OFFER` / `G729_ONLY_OFFER` fixtures. The five test
  files now `mod common;` and the drifted signatures are unified.
- media-glue: the `LINPHONE_PCMU_OFFER` / `G729_ONLY_OFFER` fixtures
  that were duplicated between `setup.rs` and `sdp_negotiation.rs`.

`acceptor_prepare.rs` keeps its own `invite` (it varies Content-Type
to exercise validation) and `controller_lifecycle.rs` its own server
variant — those are genuinely specialized, not copy-paste.
