+++ crates/media-glue/tests/sdp_negotiation.rs
@@ -26,17 +26,8 @@ fn caps_full(port: u16) -> LocalCapabilities {
     }
 }
 
-const LINPHONE_PCMU_OFFER: &str = "v=0\r\n\
-o=alice 1234 5678 IN IP4 10.0.0.5\r\n\
-s=Talk\r\n\
-c=IN IP4 10.0.0.5\r\n\
-t=0 0\r\n\
-m=audio 7078 RTP/AVP 0 8 101\r\n\
-a=rtpmap:0 PCMU/8000\r\n\
-a=rtpmap:8 PCMA/8000\r\n\
-a=rtpmap:101 telephone-event/8000\r\n\
-a=fmtp:101 0-15\r\n\
-a=sendrecv\r\n";
+mod common;
+use common::{G729_ONLY_OFFER, LINPHONE_PCMU_OFFER};
 
 const ASTERISK_OPUS_PREF_OFFER: &str = "v=0\r\n\
 o=- 1700000000 1 IN IP4 10.0.0.5\r\n\
@@ -59,15 +50,6 @@ c=IN IP4 10.0.0.5\r\n\
 t=0 0\r\n\
 m=video 5004 RTP/AVP 96\r\n\
 a=rtpmap:96 H264/90000\r\n\
-a=sendrecv\r\n";
-
-const G729_ONLY_OFFER: &str = "v=0\r\n\
-o=- 1 1 IN IP4 10.0.0.5\r\n\
-s=Talk\r\n\
-c=IN IP4 10.0.0.5\r\n\
-t=0 0\r\n\
-m=audio 7000 RTP/AVP 18\r\n\
-a=rtpmap:18 G729/8000\r\n\
 a=sendrecv\r\n";
 
 #[test]
